# ggml-cpu: replace NEON asm with intrinsics in ggml_gemv_q4_0_4x8_q8_0()